connector-v2 to log1 log2

+++ apri-sensor-redis/apri-sensor-redis.sh
@@ -30,9 +30,16 @@ journalctl --vacuum-time=1h
 # opruimen logfiles zolang winston nog niet volledig ingericht is
 # ook de oude map voorlopig meenmenen (tot ApriSensorSK2)
 cd /var/log/aprisensor
-mv SCAPE604-apri-sensor-connector-old1.log SCAPE604-apri-sensor-connector-old2.log
-mv SCAPE604-apri-sensor-connector.log SCAPE604-apri-sensor-connector-old1.log
-#cd /opt/SCAPE604/log
+if [ -f /opt/SCAPE604/git/apri-sensor/SCAPE604-apri-sensor-connector.log ]
+then
+  mv SCAPE604-apri-sensor-connector-old1.log SCAPE604-apri-sensor-connector-old2.log
+  mv SCAPE604-apri-sensor-connector.log SCAPE604-apri-sensor-connector-old1.log
+fi
+if [ -f /opt/SCAPE604/git/apri-sensor/SCAPE604-apri-sensor-connector-v.log ]
+then
+  mv SCAPE604-apri-sensor-connector-v2-old1.log SCAPE604-apri-sensor-connector-v2-old2.log
+  mv SCAPE604-apri-sensor-connector-v2.log SCAPE604-apri-sensor-connector-v2-old1.log
+fi#cd /opt/SCAPE604/log
 #mv SCAPE604-apri-sensor-connector-old1.log SCAPE604-apri-sensor-connector-old2.log
 #mv SCAPE604-apri-sensor-connector.log SCAPE604-apri-sensor-connector-old1.log
